"update App.js, Checkout, Cards and Cart
